+++ pkcs11/src/backend/key.rs
@@ -371,19 +371,28 @@ pub fn create_key_from_template(
     Ok((id, key_class, parsed.raw_id))
 }
 
-const KEYTYPE_EC_P224: ObjectIdentifier = der::oid::db::rfc5912::SECP_224_R_1;
 const KEYTYPE_EC_P256: ObjectIdentifier = der::oid::db::rfc5912::SECP_256_R_1;
 const KEYTYPE_EC_P384: ObjectIdentifier = der::oid::db::rfc5912::SECP_384_R_1;
 const KEYTYPE_EC_P521: ObjectIdentifier = der::oid::db::rfc5912::SECP_521_R_1;
 const KEYTYPE_CURVE25519: ObjectIdentifier = der::oid::db::rfc8410::ID_ED_25519;
+const KEYTYPE_EC_P256_K1: ObjectIdentifier = ObjectIdentifier::new_unwrap("1.3.132.0.10");
+const KEYTYPE_BRAINPOOL_P256: ObjectIdentifier =
+    ObjectIdentifier::new_unwrap("1.3.36.3.3.2.8.1.1.7");
+const KEYTYPE_BRAINPOOL_P384: ObjectIdentifier =
+    ObjectIdentifier::new_unwrap("1.3.36.3.3.2.8.1.1.11");
+const KEYTYPE_BRAINPOOL_P512: ObjectIdentifier =
+    ObjectIdentifier::new_unwrap("1.3.36.3.3.2.8.1.1.13");
 
 pub fn key_type_to_asn1(key_type: KeyType) -> Option<ObjectIdentifier> {
     Some(match key_type {
-        KeyType::EcP224 => KEYTYPE_EC_P224,
         KeyType::EcP256 => KEYTYPE_EC_P256,
         KeyType::EcP384 => KEYTYPE_EC_P384,
         KeyType::EcP521 => KEYTYPE_EC_P521,
         KeyType::Curve25519 => KEYTYPE_CURVE25519,
+        KeyType::EcP256K1 => KEYTYPE_EC_P256_K1,
+        KeyType::BrainpoolP256 => KEYTYPE_BRAINPOOL_P256,
+        KeyType::BrainpoolP384 => KEYTYPE_BRAINPOOL_P384,
+        KeyType::BrainpoolP512 => KEYTYPE_BRAINPOOL_P512,
         _ => return None,
     })
 }
@@ -391,11 +400,14 @@ pub fn key_type_to_asn1(key_type: KeyType) -> Option<ObjectIdentifier> {
 // returns the key size in bytes
 pub const fn key_size(t: &KeyType) -> Option<usize> {
     let size = match t {
-        KeyType::EcP224 => 224,
         KeyType::EcP256 => 256,
         KeyType::EcP384 => 384,
         KeyType::EcP521 => 521,
         KeyType::Curve25519 => 255,
+        KeyType::EcP256K1 => 256,
+        KeyType::BrainpoolP256 => 256,
+        KeyType::BrainpoolP384 => 384,
+        KeyType::BrainpoolP512 => 512,
         _ => return None,
     };
 
@@ -409,14 +421,20 @@ fn key_type_from_params(params: &[u8]) -> Option<KeyType> {
     // we can't do a match on vecs
     if oid == KEYTYPE_CURVE25519 {
         Some(KeyType::Curve25519)
-    } else if oid == KEYTYPE_EC_P224 {
-        Some(KeyType::EcP224)
     } else if oid == KEYTYPE_EC_P256 {
         Some(KeyType::EcP256)
     } else if oid == KEYTYPE_EC_P384 {
         Some(KeyType::EcP384)
     } else if oid == KEYTYPE_EC_P521 {
         Some(KeyType::EcP521)
+    } else if oid == KEYTYPE_EC_P256_K1 {
+        Some(KeyType::EcP256K1)
+    } else if oid == KEYTYPE_BRAINPOOL_P256 {
+        Some(KeyType::BrainpoolP256)
+    } else if oid == KEYTYPE_BRAINPOOL_P384 {
+        Some(KeyType::BrainpoolP384)
+    } else if oid == KEYTYPE_BRAINPOOL_P512 {
+        Some(KeyType::BrainpoolP512)
     } else {
         None
     }